Move menu container to seperate component
